Add a new component named ToDoList
